scripts/requirements.txt
google-cloud-bigquery>=3.11.0
google-cloud-functions>=1.12.0
flask>=2.3.0

